refactor: 엔티티 패키지 경로 변경에 따른 코드 수정
