+++ src/main/java/org/kakao/kakaoshopping/domain/repository/order/OrderItemRepository.java
@@ -1,6 +1,6 @@
 package org.kakao.kakaoshopping.domain.repository.order;
 
-import org.kakao.kakaoshopping.domain.entity.OrderItem;
+import org.kakao.kakaoshopping.domain.entity.order.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderItemRepository extends JpaRepository<OrderItem, Long>, OrderItemQueryRepository {
