+++ src/main/java/org/kakao/kakaoshopping/domain/repository/order/OrderQueryRepository.java
@@ -1,6 +1,6 @@
 package org.kakao.kakaoshopping.domain.repository.order;
 
-import org.kakao.kakaoshopping.domain.entity.Order;
+import org.kakao.kakaoshopping.domain.entity.order.Order;
 import org.kakao.kakaoshopping.web.common.paging.request.OrderSearchCondition;
 import org.springframework.data.domain.Page;
