+++ src/main/java/org/kakao/kakaoshopping/domain/service/order/OrderItemService.java
@@ -1,6 +1,6 @@
 package org.kakao.kakaoshopping.domain.service.order;
 
-import org.kakao.kakaoshopping.domain.entity.OrderItem;
+import org.kakao.kakaoshopping.domain.entity.order.OrderItem;
 import org.kakao.kakaoshopping.domain.repository.order.OrderItemRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
