+++ src/main/java/org/kakao/kakaoshopping/web/dto/order/response/OrderItemSimpleView.java
@@ -1,6 +1,6 @@
 package org.kakao.kakaoshopping.web.dto.order.response;
 
-import org.kakao.kakaoshopping.domain.entity.OrderItem;
+import org.kakao.kakaoshopping.domain.entity.order.OrderItem;
 
 import lombok.Data;
 
@@ -15,7 +15,7 @@ public class OrderItemSimpleView {
 
 	public OrderItemSimpleView(OrderItem orderItem) {
 		this.id = orderItem.getId();
-		this.itemId = orderItem.getItemId();
+		this.itemId = orderItem.getItem().getId();
 		this.quantity = orderItem.getQuantity();
 	}
 }
